Add general luxld linker.
This linker is the result of the combination of the existing linkers
plus a new linker for ARM. Also, MIPS shared libraries are supported now.

The driver was extensively modified to support shared libraries and also
to make use of the new linker.

The assemblers now support a 'function' extension for the 'global' directive.

musl support was improved.